Systemd socket activation fixes
Systemd sets the LISTEN_PID and LISTEN_FDS environment variables to be passed
to a service when the administrator sets up the service to be socket activated.

The application is supposed to check its current PID versus the LISTEN_PID, if
they match, the process is supposed to use the "leaked" fds listed in LISTEN_FDS.

When using docker, we want to pass this information on to the first process in the
container.  Since this process will be running as PID 1, we need to reset the LISTEN_PID
to pid 1.

Depending on the docker execdriver and whether we are in foreground mode
we need to pass/leak these FDS into the container.