devicemapper: TrimPool: Sync pool before suspending
Calling sync() on a suspended pool blocks if there were outstanding
writes, so do the sync before suspending. There is of course a race
where we may get more dirty data between the sync and the suspend,
but that should be fine. We read metadata at the suspended state
and do the discards on that before we resume it allowing the
new dirty bits to be written.

Docker-DCO-1.1-Signed-off-by: Alexander Larsson <alexl@redhat.com> (github: alexlarsson)

Docker-DCO-1.1-Signed-off-by: Alexander Larsson <alexl@redhat.com> (github: rhatdan)