Working --registry-(append|replace) with search/pull commands
Pull and Search commands iterate over a list of registries given with
options --registry-append or --registry-replace. In case of pull command
the iteration stops when a matching image is found. In case of search
command the iteration never stops - all the results are accumulated.

--registry-append is kind of confusing - it does exact oposite. It
prepends registries separated with commas in reversed order to official
index (docker.io). So the last registry specified is queried as the
first one.

TODO:
  * rename --registry-append to something like --registry-prepend.
  * add a repository origin to the results table for search command if
    multiple registries were queried

Signed-off-by: Michal Minar <miminar@redhat.com>