Generate distribution-specific Dockerfile
This PR will generate a Dockerfile specific to a distribution (RHEL, Centos and Fedora with this change but provides a pattern to follow for other distributions).  This will allow distro-specific Dockerfile to be used in build for 'make test'.  It reads from /etc/redhat-release and generates a new Dockerfile for that distribution.

Signed-off-by: Sally O'Malley <somalley@redhat.com>

Docker-DCO-1.1-Signed-off-by: Regan McCooey rmccooey27@aol.com (github:
rhatdan)