Confirm a push to public Docker registry
Ask the user to confirm a push of image to public Docker registry. The
confirmation can be suppressed with the use of `-f, --force` flag.

Added --confirm-def-push boolean flag to Docker daemon. It causes
client to ask for confirmation before a push to public registry. It
defaults to true. If set to false, push won't be restricted in any
way unless the registry in question is blocked.

No confirmation is needed for push to unofficial registry.

Signed-off-by: Michal Minar <miminar@redhat.com>

Docker-DCO-1.1-Signed-off-by: Dan Walsh <dwalsh@redhat.com> (github: rhatdan)

Dan Walsh <dwalsh@redhat.com>