Add dockerhooks program to run hooks under runc
With the addition of runc/hooks support we want to add a feature
to allow third parties to run helper programs before a docker container
gets started and just after the container finishes.

One example of this is a package we want to add RegisterMachine.

An example of a dockerhook might be for people wanting to do better logging
of starting and stopping of containers.  For example have a log agent that
records when a container starts and stops and then sends a message to a
monitoring station.

Dockerhooks reads an environment variable DOCKER_HOOKS_PATH for a directory
to search for hooks, if the directory does not exist dockerhooks just exits.

docker daemon uses the search mechanism for dockerinit to look for dockerhooks
it also sets the DOCKER_HOOKS_PATH based off the directory containers dockerhooks.

and will execute them via runc/libcontainer using PreStart and PostStop.

Signed-off-by: Sally O'Malley <somalley@redhat.com>

Signed-off-by: Dan Walsh <dwalsh@redhat.com>