Discard the nil result when getpwuid() fails
After warning that getpwuid() failed to resolve a UID to a passwd
structure, return an empty name and the error rather than attempting to
read an element from the returned nil structure.  Fixes RHBZ#1275593.

Also, don't bother looking up the name of a user for a given login UID
when that UID is -1, as that's the value we get back when none has been
set.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com> (github: nalind)