Fix some exec-related crashes in integration tests
If we fail at any step of retrieving the client's login UID in
getLoginUid(), return an error immediately after logging it, instead of
proceeding.
We're building dynamic binaries now, so os/user works, so we don't need
to directly call into getpwuid() ourselves to avoid hitting a stub
implementation.
During "docker exec", the container ID may not be resolvable, so check
for errors in attempting to read the container's configuration.
Only close a syslog connection if we were successful in opening it.