Add mountlabel to diff drivers
We should be mounting the images with the mount label when doing a
commit operation.  We have seen a leaked mount point cause the commit
fail on SELinux systems.  If the leaked mount point is a context mount
and the diff tries to mount the image without the context mount, then
the kernel denies the mount.

Signed-off-by: Daniel J Walsh <dwalsh@redhat.com>